requirements.txt
awesome-slugify
google-api-python-client
progressbar2
oauth2client
-e ./youtube-upload

# for gen_cal.py
python-dateutil
icalendar
pytz
